+++ piyavkin.anton/S1/output.hpp
@@ -6,7 +6,7 @@
 
 namespace piyavkin
 {
-  using pair_t = std::pair< std::string, piyavkin::List< unsigned int >* >*;
+  using pair_t = std::pair< std::string, piyavkin::List< unsigned long long >* >*;
   void output(std::ostream& out, const pair_t pairs, size_t size, size_t max_size_list);
 }
 #endif
